Extract scan specific logic out of bittrex exchange api module
