+++ lib/exchanges/bittrex.js
@@ -1,15 +1,11 @@
 const Promise = require('bluebird');
 const moment = require('moment');
-const _ = require('lodash');
-const fs = Promise.promisifyAll(require('fs'));
 const parseNum = require('parse-decimal-number');
 
 const bittrexAPI = require('@you21979/bittrex.com');
 const keys = require('../../api-key')['bittrex'];
 const bittrex = bittrexAPI.createPrivateApi(keys.api_key, keys.api_secret, 'I am Bot');
 const bittrexPub = bittrexAPI.PublicApi;
-
-const history = require('../../local/bittrex-history');
 
 module.exports = {
 	/* Sample output
@@ -247,10 +243,6 @@ module.exports = {
 		});
 	},
 
-	getMarkets: () => {
-		return bittrexPub.getMarkets()
-	},
-
 	getOldOrderHistory: () => {
 		/*
 		const CSVReader = require('promised-csv');
@@ -354,194 +346,20 @@ module.exports = {
 			});
 	},
 
-	getMarketHistory: (coin, count, currency) => {
+	getMarkets: () => {
+		return bittrexPub.getMarkets();
+	},
+
+	getMarketHistory: (coin, maxCount, currency) => {
 		currency = (currency || 'BTC').toUpperCase();
 		let market = `${currency}-${coin.toUpperCase()}`;
-		return bittrexPub.getMarketHistory(market, count);
+		return bittrexPub.getMarketHistory(market, maxCount);
 	},
 
 	getMarketSummaries: () => {
 		return bittrexPub.getMarketSummaries();
-	},
-
-	findMarketMovement: () => {
-		return Promise.all([
-			module.exports.getMarketSummaries(),
-			module.exports.updateHistory(history)
-		])
-			.then(res => {
-				const tickers = res[0];
-				const updated = res[1];
-				let numUpdates = 0;
-				tickers.forEach(ticker => {
-					if (ticker["MarketName"].indexOf('BTC') === 0) {
-						const coin = ticker["MarketName"].split('-')[1];
-						if (updated[coin] && updated[coin].trades.length) {
-							const lastTrade = updated[coin].trades[0];
-							const dtTicker = moment(ticker["TimeStamp"] + 'Z');
-							const dtLast = moment(lastTrade.ts);
-							const diff = dtTicker.diff(moment(lastTrade.ts), 'minutes');
-							if (diff >= 2) {
-								//console.log(`${diff} minutes since last trade for ${coin}, adding to history`);
-								updated[coin].trades.unshift({
-									ts: +dtTicker,
-									price: ticker["Last"]
-								});
-								numUpdates++;
-							}
-						}
-					}
-				});
-				console.log(`Updated ${numUpdates} coins`);
-				findBigChanges(updated);
-				const toSave = cleanupTradeHistory(updated);
-				return module.exports.saveHistory(toSave)
-					.then(() => {
-						return updated;
-					});
-			});
-	},
-
-	updateHistory: (currHistory) => {
-		return module.exports.getMarkets()
-			.then((markets) => {
-				const toAdd = [];
-				markets.forEach(mkt => {
-					if (mkt['BaseCurrency'] === 'BTC' && !currHistory[mkt['MarketCurrency']]) {
-						toAdd.push(mkt['MarketCurrency']);
-					}
-				});
-				if (toAdd.length) {
-					const promises = [];
-					for (let i = 0; i < 40 && i < (toAdd.length - 1); i++) {
-						const currCoin = toAdd[i];
-						console.log(`Fetching history for ${currCoin}`);
-						promises.push(module.exports.getMarketHistory(currCoin, 300)
-							.then(trades => {
-								return {
-									coin: currCoin,
-									trades: trades
-								};
-							}));
-					}
-					return Promise.all(promises)
-						.then(queries => {
-							queries.forEach(tradeInfo => {
-								const trades = reduceTradeHistory(tradeInfo.trades);
-								currHistory[tradeInfo.coin] = {
-									coin: tradeInfo.coin,
-									trades: trades
-								};
-							});
-							return currHistory;
-						});
-				}
-				return currHistory;
-			});
-	},
-
-	saveHistory: (toSave) => {
-		const fileStr = `module.exports = ${JSON.stringify(toSave, null, 2)};`;
-		return fs.writeFileAsync('./local/bittrex-history.js', fileStr, 'utf8');
 	}
 };
-
-const reduceTradeHistory = (rawTradeData) => {
-	const trades = rawTradeData.map(data => {
-		const dt = moment(data.TimeStamp + 'Z');
-		return {
-			ts: +dt,
-			price: data.Price
-		};
-	});
-	trades.sort((a, b) => {
-		return b.ts - a.ts;
-	});
-
-	const origCount = trades.length;
-	let lastTrade = trades[0];
-	let dtLast = moment(lastTrade.ts);
-	const reduced = [lastTrade];
-
-	trades.forEach(trade => {
-		let dt = moment(trade.ts);
-		if (dtLast.diff(dt, 'minutes') >= 2) {
-			lastTrade = trade;
-			dtLast = dt;
-			reduced.push(lastTrade);
-		}
-	});
-
-	return reduced;
-};
-
-const cleanupTradeHistory = (market) => {
-	const clean = {};
-	Object.keys(market).forEach(key => {
-		const data = market[key];
-		if (data.trades.length) {
-			const dtLast = moment(data.trades[0].ts);
-			const newTrades = data.trades.filter(trade => {
-				return dtLast.diff(moment(trade.ts), 'minutes') <= 120;
-			});
-			clean[key] = {
-				coin: data.coin,
-				trades: newTrades
-			};
-		}
-	});
-	return clean;
-};
-
-const findBigChanges = (market) => {
-	const swings = [];
-
-	Object.keys(market).forEach(coin => {
-		let trades = market[coin].trades;
-		if (trades.length) {
-			let lastPrice = trades[0].price;
-			let dtLast = moment(trades[0].ts);
-			trades.some(trade => {
-				let dtTrade = moment(trade.ts);
-				let mins = dtLast.diff(dtTrade, 'minutes');
-				if (mins > 0 && mins < 60) {
-					let chg = lastPrice - trade.price;
-					let pctChg = _.floor(((chg / trade.price) * 100), 2);
-					let absChg = Math.abs(pctChg);
-					if (absChg >= 6 || (absChg >= 4 && mins < 30) || (absChg >= 2 && mins <= 15)) {
-						swings.push({
-							coin: coin,
-							change: pctChg,
-							minutes: mins
-						});
-						return true;
-					}
-				}
-				return false;
-			});
-		}
-	});
-
-	swings.sort((a,b) => {
-		return b.change - a.change;
-	});
-
-	swings.forEach(next => {
-		console.log(`${next.change}% in ${next.minutes} mins (${next.coin})`);
-	});
-
-	return swings;
-};
-
-/*
-return module.exports.findMarketMovement()
-	.then((updated) => {
-		Object.keys(updated).forEach(coin => {
-			const data = updated[coin];
-		});
-		process.exit();
-	});
-*/
 
 /* Sample Orders
 {
